+++ src/main/java/org/morphix/convert/strategy/FinderMapKeyStrategy.java
@@ -15,9 +15,12 @@ package org.morphix.convert.strategy;
 import static org.morphix.reflection.ExtendedField.of;
 
 import java.lang.reflect.Field;
+import java.rmi.Naming;
 import java.util.List;
 import java.util.Map;
 
+import org.morphix.convert.function.SimpleConverter;
+import org.morphix.lang.Case;
 import org.morphix.reflection.ExtendedField;
 
 /**
@@ -26,6 +29,16 @@ import org.morphix.reflection.ExtendedField;
  * @author Radu Sebastian LAZIN
  */
 public class FinderMapKeyStrategy implements FieldFinderStrategy {
+
+	/**
+	 * List of key converters to apply to the source field name when searching for a value in the map. The converters are
+	 * applied in order, and the search stops as soon as a value is found for a converted key. The converters include:
+	 * <ul>
+	 * <li>{@link Naming#kebabCase(String)}</li>
+	 * </ul>
+	 */
+	private static final List<SimpleConverter<String, String>> KEY_CONVERTERS = List.of(
+			Case.KEBAB::convert);
 
 	/**
 	 * Default constructor.
@@ -40,8 +53,30 @@ public class FinderMapKeyStrategy implements FieldFinderStrategy {
 	@Override
 	public <T> ExtendedField find(final T source, final List<ExtendedField> fields, final String sourceFieldName) {
 		if (source instanceof Map<?, ?> sourceMap) {
-			return of((Field) null, sourceMap.get(sourceFieldName));
+			Object value = sourceMap.get(sourceFieldName);
+			if (null == value) {
+				value = findValueByKeyConverters(sourceMap, sourceFieldName);
+			}
+			return of((Field) null, value);
 		}
 		return ExtendedField.EMPTY;
+	}
+
+	/**
+	 * Finds a value in the given map by applying the key converters to the source field name.
+	 *
+	 * @param sourceMap the map to search for the value
+	 * @param sourceFieldName the source field name to convert and search for
+	 * @return the value found in the map, or {@code null} if no value is found
+	 */
+	private static Object findValueByKeyConverters(final Map<?, ?> sourceMap, final String sourceFieldName) {
+		for (SimpleConverter<String, String> keyConverter : KEY_CONVERTERS) {
+			String convertedKey = keyConverter.convert(sourceFieldName);
+			Object value = sourceMap.get(convertedKey);
+			if (null != value) {
+				return value;
+			}
+		}
+		return null;
 	}
 }